split process into two classes
