+++ common/process/CMakeLists.txt
@@ -22,11 +22,14 @@ add_definitions(${spdlog_DEFINITIONS})
 
 if(${CMAKE_SYSTEM_NAME} STREQUAL "Linux")
   set(SOURCES
+    src/child_process.cc
+    #src/fork.cc
     src/process_args.cc
     src/process.cc
     src/spawnp_launcher.cc)
 else()
   set(SOURCES
+    src/child_process.cc
     src/process_args.cc
     src/process.cc)
 endif()
